Move CLI to the separate file
